Split mode profile builders by concern
